% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Populationpack2016-package.R
\docType{package}
\name{Populationpack2016-package}
\alias{Populationpack2016}
\alias{Populationpack2016-package}
\title{Populationpack2016: Information pertained to Mashhad's population, IRAN.}
\description{
The goal of Populationpack2016 is to analyze information about population based on number of families, owners, tenants and residential units of all neighbourhoods in Mashhad, Iran.

The goal of Populationpack2016 is to analyze information about population based on number of families, owners, tenants and residential units of all neighbourhoods in Mashhad, Iran.
}
\references{
\url{https://github.com/NiloofarNL/Populationpack2016}
}
\author{
\strong{Maintainer}: Niloofar Nalaee \email{nalaeen@mcmaster.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


Niloofar Nalaee, School of Earth, Environment and Society, McMaster University \email{nalaeen@mcmaster.ca}
}
\keyword{internal}
